Update README.md
Hoping that this gets updated by the time it's handed in, having trouble getting the file uploaded so may have to settle for a project declaration showing my work.. unideal.

A README has been solidly added to reflect the work put into building the project, wasn't entirely sure about how specific I needed to be on this markdown, so in avoidance of showing too much of the coding (as you can access it here in the zip of course), I've just gone through a step-by-step of my thought process.

Hopefully it's clear enough to follow, as time constraints really have beaten me for this one. Unfortunate.

At least it's thorough! Have a nice read ^^